requirements.txt
pip~=25.3
typing_extensions~=4.15.0
attrs~=25.3.0
slack_sdk~=3.41.0
pyslack~=0.5.0
aiohttp~=3.11.18
websockets~=15.0.1
slack_bolt~=1.27.0
Flask~=3.1.0
python-dotenv~=1.1.0
APScheduler~=3.11.2